Fix ambito-grid text overflow. Add materias.html
Fixed a rpoblem on the Ambitos grid where if the names of the ambitos and materias where kinda long it would create a text overflow issue.
It was fixed by removing fixed heights from the flippable cards. The order of the corresponding javascripts functions had to be changed so the cards load first then the masonry grid system.
Created the materias.html view that shows teh sections fo the Materia and the ODAs in their respective positions.
ODAS are rendered in a hidden area and then auto-positioned into where they belong according to the 'data-n' field.
ODAs also have 1 of 4 states:
Unopened
In-progress
Completed
Finalized.
A small CSS driven animation is shown when hovered over the ODA corresponding to the state it is in.